+++ packages/components/src/components/input/input.component.ts
@@ -1,5 +1,5 @@
 import { CSSResult, html, nothing, PropertyValueMap } from 'lit';
-import { property, state } from 'lit/decorators.js';
+import { property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 
@@ -10,6 +10,7 @@ import { DataAriaLabelMixin } from '../../utils/mixins/DataAriaLabelMixin';
 import { FormInternalsMixin, AssociatedFormControl } from '../../utils/mixins/FormInternalsMixin';
 import { AutoFocusOnMountMixin } from '../../utils/mixins/AutoFocusOnMountMixin';
 import { KeyToActionMixin, ACTIONS, NAV_MODES } from '../../utils/mixins/KeyToActionMixin';
+import { CharacterLimitMixin } from '../../utils/mixins/CharacterLimitMixin';
 
 import type { AutoCapitalizeType, AutoCompleteType, InputType } from './input.types';
 import { AUTO_CAPITALIZE, AUTO_COMPLETE, DEFAULTS, PREFIX_TEXT_OPTIONS } from './input.constants';
@@ -66,7 +67,6 @@ import styles from './input.styles';
  * @cssproperty --mdc-input-support-text-color - Text color for the help text
  * @cssproperty --mdc-input-selection-text-color - Text color for the selected text
  * @cssproperty --mdc-input-selection-background-color - Background color for the selected text
- * @cssproperty --mdc-input-text-secondary-normal - Text color for the character counter
  *
  * @csspart label - The label element.
  * @csspart label-text - The container for the label and required indicator elements.
@@ -87,7 +87,9 @@ import styles from './input.styles';
  */
 
 class Input
-  extends KeyToActionMixin(AutoFocusOnMountMixin(FormInternalsMixin(DataAriaLabelMixin(FormfieldWrapper))))
+  extends CharacterLimitMixin(
+    KeyToActionMixin(AutoFocusOnMountMixin(FormInternalsMixin(DataAriaLabelMixin(FormfieldWrapper)))),
+  )
   implements AssociatedFormControl
 {
   /**
@@ -173,30 +175,6 @@ class Input
   @property({ type: String, reflect: true, attribute: 'data-aria-describedby' })
   dataAriaDescribedby: string | null = null;
 
-  /**
-   * The maximum character limit for the input field for character counter.
-   */
-  @property({ type: Number, attribute: 'max-character-limit' }) maxCharacterLimit?: number;
-
-  /**
-   * Template string for the announcement that will be read by screen readers when the max character limit is set.
-   * Consumers must use the placeholders `%{number-of-characters}` and `%{max-character-limit}` in the string,
-   * which will be dynamically replaced with the actual values at runtime.
-   * For example: `%{number-of-characters} out of %{max-character-limit} characters are typed.`
-   * Example output: "93 out of 140 characters are typed."
-   */
-  @property({ type: String, attribute: 'character-limit-announcement' }) characterLimitAnnouncement?: string;
-
-  /**
-   * @internal
-   */
-  @state() private ariaLiveAnnouncer?: string;
-
-  /**
-   * @internal
-   */
-  private characterLimitExceedingFired: boolean = false;
-
   override connectedCallback(): void {
     super.connectedCallback();
     this.updateComplete
@@ -281,65 +259,7 @@ class Input
       this.handleCharacterOverflowStateChange();
     }
     if (changedProperties.has('helpText')) {
-      this.announceMaxLengthWarning();
-    }
-  }
-
-  /**
-   * Dispatches the character overflow state change event.
-   * @returns void
-   */
-  private dispatchCharacterOverflowStateChangeEvent() {
-    this.dispatchEvent(
-      new CustomEvent('limitexceeded', {
-        detail: {
-          currentCharacterCount: this.value.length,
-          maxCharacterLimit: this.maxCharacterLimit,
-          value: this.value,
-        },
-        bubbles: true,
-        composed: true,
-      }),
-    );
-  }
-
-  /**
-   * Handles the character overflow state change.
-   * Dispatches the character overflow state change event if the character limit is exceeded or restored.
-   * @returns void
-   */
-  private handleCharacterOverflowStateChange() {
-    if (this.maxCharacterLimit) {
-      if (this.value.length > this.maxCharacterLimit && !this.characterLimitExceedingFired) {
-        this.dispatchCharacterOverflowStateChangeEvent();
-        this.characterLimitExceedingFired = true;
-      } else if (this.value.length <= this.maxCharacterLimit && this.characterLimitExceedingFired) {
-        this.dispatchCharacterOverflowStateChangeEvent();
-        this.characterLimitExceedingFired = false;
-      }
-    }
-  }
-
-  /**
-   * Announces the character limit warning based on the current value length.
-   * If the value length exceeds the max character limit, the help text is announced (if help text is present).
-   * If the value length does not exceed the max character limit, then the character limit announcement is announced.
-   */
-  private announceMaxLengthWarning() {
-    this.ariaLiveAnnouncer = '';
-    if (!this.maxCharacterLimit || this.value.length === 0) {
-      return;
-    }
-    if (this.helpText && this.value.length > this.maxCharacterLimit) {
-      this.updateComplete
-        .then(() => {
-          this.ariaLiveAnnouncer = this.helpText;
-        })
-        .catch(() => {});
-    } else if (this.characterLimitAnnouncement && this.value.length <= this.maxCharacterLimit) {
-      this.ariaLiveAnnouncer = this.characterLimitAnnouncement
-        .replace('%{number-of-characters}', this.value.length.toString())
-        .replace('%{max-character-limit}', this.maxCharacterLimit.toString());
+      this.announceCharacterLimitWarning();
     }
   }
 
@@ -351,7 +271,7 @@ class Input
   private updateValue() {
     this.value = this.inputElement.value;
     this.internals.setFormValue(this.inputElement.value);
-    this.announceMaxLengthWarning();
+    this.announceCharacterLimitWarning();
   }
 
   /**
@@ -513,17 +433,6 @@ class Input
     />`;
   }
 
-  protected renderCharacterCounter() {
-    if (!this.maxCharacterLimit) {
-      return nothing;
-    }
-    return html`
-      <mdc-text part="character-counter" tagname="span" type=${DEFAULTS.CHARACTER_COUNTER_TYPE}>
-        ${this.value.length}/${this.maxCharacterLimit}
-      </mdc-text>
-    `;
-  }
-
   protected renderInputFooter() {
     if (!this.helpText && !this.maxCharacterLimit) {
       return nothing;
@@ -544,7 +453,7 @@ class Input
       </div>
       <mdc-screenreaderannouncer
         identity="${this.inputId}"
-        announcement="${ifDefined(this.ariaLiveAnnouncer)}"
+        announcement="${ifDefined(this.characterLimitAriaLiveAnnouncer)}"
         data-aria-live="polite"
         delay="500"
       ></mdc-screenreaderannouncer>
