+++ packages/components/src/components/input/input.styles.ts
@@ -12,7 +12,6 @@ const styles = [
       --mdc-input-selection-text-color: var(--mds-color-theme-inverted-text-primary-normal);
       --mdc-input-selection-background-color: var(--mds-color-theme-background-accent-active);
       --mdc-input-height: 2rem;
-      --mdc-input-text-secondary-normal: var(--mds-color-theme-text-secondary-normal);
     }
     :host,
     :host::part(input-container),
@@ -135,7 +134,7 @@ const styles = [
 
     :host::part(character-counter) {
       margin-left: auto;
-      color: var(--mdc-input-text-secondary-normal);
+      color: var(--mdc-input-support-text-color);
     }
 
     :host([help-text-type='error'])::part(character-counter) {
