+++ packages/components/src/components/textarea/textarea.component.ts
@@ -1,6 +1,6 @@
 import { CSSResult, html, nothing, PropertyValueMap } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import { property, query, state } from 'lit/decorators.js';
+import { property, query } from 'lit/decorators.js';
 
 import FormfieldWrapper from '../formfieldwrapper';
 import { DEFAULTS as FORMFIELD_DEFAULTS, VALIDATION } from '../formfieldwrapper/formfieldwrapper.constants';
@@ -11,6 +11,7 @@ import { FormInternalsMixin } from '../../utils/mixins/FormInternalsMixin';
 import { AutoFocusOnMountMixin } from '../../utils/mixins/AutoFocusOnMountMixin';
 import { ACTIONS, KeyToActionMixin } from '../../utils/mixins/KeyToActionMixin';
 import { KeyDownHandledMixin } from '../../utils/mixins/KeyDownHandledMixin';
+import { CharacterLimitMixin } from '../../utils/mixins/CharacterLimitMixin';
 
 import { AUTO_COMPLETE, WRAP, DEFAULTS } from './textarea.constants';
 import type { WrapType, AutoCompleteType } from './textarea.types';
@@ -102,8 +103,10 @@ import styles from './textarea.styles';
  * @cssproperty --mdc-textarea-container-background-color - Background color for the textarea container
  */
 
-class Textarea extends KeyDownHandledMixin(
-  KeyToActionMixin(AutoFocusOnMountMixin(FormInternalsMixin(DataAriaLabelMixin(FormfieldWrapper)))),
+class Textarea extends CharacterLimitMixin(
+  KeyDownHandledMixin(
+    KeyToActionMixin(AutoFocusOnMountMixin(FormInternalsMixin(DataAriaLabelMixin(FormfieldWrapper)))),
+  ),
 ) {
   /**
    * The placeholder text that is displayed when the textarea field is empty.
@@ -156,20 +159,6 @@ class Textarea extends KeyDownHandledMixin(
   @property({ type: Number }) minlength?: number;
 
   /**
-   * The maximum character limit for the textarea field for character counter.
-   */
-  @property({ type: Number, attribute: 'max-character-limit' }) maxCharacterLimit?: number;
-
-  /**
-   * Template string for the announcement that will be read by screen readers when the max character limit is set.
-   * Consumers must use the placeholders `%{number-of-characters}` and `%{max-character-limit}` in the string,
-   * which will be dynamically replaced with the actual values at runtime.
-   * For example: `%{number-of-characters} out of %{max-character-limit} characters are typed.`
-   * Example output: "93 out of 140 characters are typed."
-   */
-  @property({ type: String, attribute: 'character-limit-announcement' }) characterLimitAnnouncement?: string;
-
-  /**
    * Controls whether the textarea is resizable via the resize button.
    * When set to false, the resize button will be hidden.
    * @default false
@@ -188,12 +177,6 @@ class Textarea extends KeyDownHandledMixin(
    * The textarea element
    */
   @query('textarea') override inputElement!: HTMLTextAreaElement;
-
-  /** @internal */
-  @state() private ariaLiveAnnouncer?: string;
-
-  /** @internal */
-  private characterLimitExceedingFired: boolean = false;
 
   /** @internal */
   private resizeStartY: number = 0;
@@ -290,42 +273,7 @@ class Textarea extends KeyDownHandledMixin(
     }
     // When helpText gets changed, we need to re-announce the max length warning
     if (changedProperties.has('helpText')) {
-      this.announceMaxLengthWarning();
-    }
-  }
-
-  /**
-   * Dispatches the character overflow state change event.
-   * @returns void
-   */
-  private dispatchCharacterOverflowStateChangeEvent() {
-    this.dispatchEvent(
-      new CustomEvent('limitexceeded', {
-        detail: {
-          currentCharacterCount: this.value.length,
-          maxCharacterLimit: this.maxCharacterLimit,
-          value: this.value,
-        },
-        bubbles: true,
-        composed: true,
-      }),
-    );
-  }
-
-  /**
-   * Handles the character overflow state change.
-   * Dispatches the character overflow state change event if the character limit is exceeded or restored.
-   * @returns void
-   */
-  private handleCharacterOverflowStateChange() {
-    if (this.maxCharacterLimit) {
-      if (this.value.length > this.maxCharacterLimit && !this.characterLimitExceedingFired) {
-        this.dispatchCharacterOverflowStateChangeEvent();
-        this.characterLimitExceedingFired = true;
-      } else if (this.value.length <= this.maxCharacterLimit && this.characterLimitExceedingFired) {
-        this.dispatchCharacterOverflowStateChangeEvent();
-        this.characterLimitExceedingFired = false;
-      }
+      this.announceCharacterLimitWarning();
     }
   }
 
@@ -337,33 +285,7 @@ class Textarea extends KeyDownHandledMixin(
   private updateValue() {
     this.value = this.textarea.value;
     this.internals.setFormValue(this.textarea.value);
-    this.announceMaxLengthWarning();
-  }
-
-  /**
-   * Announces the character limit warning based on the current value length.
-   * If the value length exceeds the max character limit, the help text is announced (if help text is present).
-   * If the value length does not exceed the max character limit, then the character limit announcement is announced.
-   */
-  private announceMaxLengthWarning() {
-    this.ariaLiveAnnouncer = '';
-    if (!this.maxCharacterLimit || this.value.length === 0) {
-      return;
-    }
-    if (this.helpText && this.value.length > this.maxCharacterLimit) {
-      // We need to assign the same value multiple times, when the input reaches the max limit,
-      // Lit does a `===` strict comparison and doesn't update the value
-      // Hence we need to manually wait for the update to complete and then assign the value.
-      this.updateComplete
-        .then(() => {
-          this.ariaLiveAnnouncer = this.helpText;
-        })
-        .catch(() => {});
-    } else if (this.characterLimitAnnouncement && this.value.length <= this.maxCharacterLimit) {
-      this.ariaLiveAnnouncer = this.characterLimitAnnouncement
-        .replace('%{number-of-characters}', this.value.length.toString())
-        .replace('%{max-character-limit}', this.maxCharacterLimit.toString());
-    }
+    this.announceCharacterLimitWarning();
   }
 
   /**
@@ -380,17 +302,6 @@ class Textarea extends KeyDownHandledMixin(
     this.updateValue();
     const EventConstructor = event.constructor as typeof Event;
     this.dispatchEvent(new EventConstructor(event.type, event));
-  }
-
-  protected renderCharacterCounter() {
-    if (!this.maxCharacterLimit) {
-      return nothing;
-    }
-    return html`
-      <mdc-text part="character-counter" tagname="span" type=${DEFAULTS.CHARACTER_COUNTER_TYPE}>
-        ${this.value.length}/${this.maxCharacterLimit}
-      </mdc-text>
-    `;
   }
 
   protected renderTextareaFooter() {
@@ -512,7 +423,7 @@ class Textarea extends KeyDownHandledMixin(
         ></textarea>
         <mdc-screenreaderannouncer
           identity="${this.inputId}"
-          announcement="${ifDefined(this.ariaLiveAnnouncer)}"
+          announcement="${ifDefined(this.characterLimitAriaLiveAnnouncer)}"
           data-aria-live="polite"
           delay="500"
         ></mdc-screenreaderannouncer>
